requirements.txt
Django>=4.2.6,<4.3
djangorestframework>=3.14.0,<3.15
psycopg2>=2.9.9,<2.10
drf-spectacular>=0.26.5,<0.27
django-phonenumber-field[phonenumberslite]>=7.2.0,<7.3
